Release SolSharp 2.0.0 — client parity and hardening
 Release SolSharp 2.0.0 — client parity and hardening (#1)

  Deliver the SolSharp 2.0 major release against pinned Anza Solana SDK,
  Agave, and SPL client contracts.

  Protocol and transactions
  - Add the fixed-width Hash primitive and typed blockhash APIs.
  - Add SIMD-0385 Message V1, TransactionConfigV1, wire routing,
    validation, compilation, decompilation, signing, and serialization.
  - Complete Legacy, V0, and V1 transaction round-trips and malformed-wire
    validation.
  - Expand transaction builders, offline signing, durable nonce, and
    address lookup table support.

  Wallet and cryptography
  - Harden Ed25519 key generation, import, export, signing, disposal, and
    zeroization.
  - Add domain-separated Solana off-chain messages.
  - Add BLS12-381 keypairs, signatures, proofs of possession, vote-account
    binding, and PoP-gated same-message aggregation.
  - Add secure binary, Base64, and JSON key round-trips.
  - Validate cryptographic behavior through managed and Native AOT smoke
    tests.

  Native and SPL programs
  - Add Stake, Vote, legacy and upgradeable BPF loaders, Loader V4,
    Feature Gate, Address Lookup Table state, and raw Memo support.
  - Add Ed25519, Secp256k1, and Secp256r1 precompile builders and offset
    codecs.
  - Expand System Program helpers and account-state decoders.
  - Complete classic Token, Token-2022, and Associated Token Account client
    surfaces.
  - Add token metadata, token group, transfer hook, confidential transfer,
    confidential fees, confidential mint/burn, permissioned burn, and
    ElGamal registry APIs.
  - Add strict token instruction, extension, TLV, proof-location, and
    account-state decoders.
  - Add upstream-compatible account metas, discriminators, POD layouts,
    and known-answer vectors.

  RPC and WebSocket
  - Cover the applicable Agave HTTP RPC method and PubSub family surface.
  - Add typed configuration unions, account encodings, filters, context
    responses, data slices, transaction details, and version handling.
  - Add exact base58, base64, base64+zstd, and jsonParsed account-data
    models.
  - Add strict response-envelope, account, block, transaction, token,
    simulation, vote, and cluster-state validation.
  - Harden subscription routing, signature notification lifecycle,
    reconnect behavior, cancellation, and malformed-notification
    isolation.
  - Add Address Lookup Table lifecycle, cooldown, context-slot, and active
    address semantics.

  Core data and state
  - Add current sysvar IDs and Clock, Rent, Epoch Schedule, Epoch Rewards,
    Last Restart Slot, Slot Hashes, Slot History, and Stake History
    decoders.
  - Add strict Stake, Vote, loader, feature, token, and lookup-table state
    parsing with hostile-input bounds.
  - Harden Borsh, short-vector, Base58, public-key, hash, and JSON codecs.

  Quality and release engineering
  - Add 1,698 passing automated tests backed by pinned upstream vectors.
  - Establish a reproducible 93.7% line-coverage baseline with generated
    sources excluded.
  - Add zero-warning build and analyzer gates.
  - Validate all functional assemblies through a packed-package Native AOT
    consumer.
  - Add package validation against SolSharp 1.3.0 with reviewed 2.0
    compatibility suppressions.
  - Add direct and transitive dependency audits, dependency review,
    CodeQL security-extended analysis, OpenSSF Scorecard, Dependabot, and
    package provenance attestations.
  - Pin GitHub Actions to immutable revisions and harden release endpoint,
    package-source, SDK, and integration-test validation.

  Documentation
  - Rewrite README and NuGet documentation for the 2.0 surface.
  - Add runnable usage examples, security guidance, migration notes,
    third-party notices, and the pinned Rust parity matrix.
  - Document the comparison with published Solnet 8.7.0 and the official
    Rust contracts.
  - Update changelog, package metadata, security policy, and release
    provenance links.

  BREAKING CHANGE:
  - Introduce deliberate 2.0 RPC model and wire-width corrections,
    including typed transaction versions and closed response unions.
  - Tighten malformed-input handling and nullability contracts.
  - Typed Hash overloads can make older calls using untyped null or default
    literals ambiguous.
  - Existing 1.x consumers should recompile and follow the migration notes
    in CHANGELOG.md.

+++ src/SolSharp.Rpc/DataSlice.cs
@@ -11,6 +11,6 @@ namespace SolSharp.Rpc;
 /// <param name="Length">The number of bytes to return.</param>
 public sealed record DataSlice(
     [property: JsonPropertyName("offset")]
-    int Offset,
+    ulong Offset,
     [property: JsonPropertyName("length")]
-    int Length);
+    ulong Length);
